refactor: sort and rename lib
