+++ packages/sshnoports/lib/sshnp/cleanup.dart
@@ -1,7 +1,7 @@
 // dart packages
 import 'dart:io';
 // local packages
-import 'package:sshnoports/utils.dart';
+import 'package:sshnoports/shared/utils.dart';
 // atPlatform packages
 import 'package:at_utils/at_logger.dart';
 
@@ -21,15 +21,16 @@ Future<void> cleanUp(String sessionId, AtSignLogger logger) async {
   await removeSession(sshHomeDirectory, sessionId, logger);
 }
 
-Future<int> deleteFile(String fileName, AtSignLogger logger) async {
+Future<bool> deleteFile(String fileName, AtSignLogger logger) async {
   try {
     final file = File(fileName);
 
     await file.delete();
+    return true;
   } catch (e) {
     logger.severe("Error deleting file : $fileName");
+    return false;
   }
-  return 0;
 }
 
 Future<void> removeSession(
